change ttl uint32_t to uint64_t

+++ src/tablet/tablet_impl.cc
@@ -649,7 +649,7 @@ void TabletImpl::LoadTable(RpcController* controller,
     }
     uint32_t tid = request->tid();
     uint32_t pid = request->pid();
-    uint32_t ttl = request->ttl();
+    uint64_t ttl = request->ttl();
     std::string name = request->name();
     uint32_t seg_cnt = 8;
     if (request->seg_cnt() > 0 && request->seg_cnt() < 32) {
@@ -680,11 +680,11 @@ void TabletImpl::LoadTable(RpcController* controller,
         }       
     }
     done->Run();
-    LOG(INFO, "create table with id %d pid %d name %s seg_cnt %d ttl %d", tid, 
+    LOG(INFO, "create table with id %d pid %d name %s seg_cnt %d ttl %lu", tid, 
             pid, name.c_str(), seg_cnt, ttl);
     
     // load snapshot data
-    Table* table = GetTable(tid, pid);        
+    Table* table = GetTable(tid, pid);
     if (table == NULL) {
         LOG(WARNING, "table with tid %ld and pid %ld does not exist", tid, pid);
         return; 
@@ -713,7 +713,7 @@ void TabletImpl::LoadTable(RpcController* controller,
     table->UnRef();
     if (ttl > 0) {
         gc_pool_.DelayTask(FLAGS_gc_interval * 60 * 1000, boost::bind(&TabletImpl::GcTable, this, tid, pid));
-        LOG(INFO, "table %s with tid %ld pid %ld enable ttl %ld", name.c_str(), tid, pid, ttl);
+        LOG(INFO, "table %s with tid %ld pid %ld enable ttl %lu", name.c_str(), tid, pid, ttl);
     }
 }
 
@@ -787,7 +787,7 @@ void TabletImpl::CreateTable(RpcController* controller,
     }
     uint32_t tid = request->tid();
     uint32_t pid = request->pid();
-    uint32_t ttl = request->ttl();
+    uint64_t ttl = request->ttl();
     std::string name = request->name();
     uint32_t seg_cnt = 8;
     if (request->seg_cnt() > 0 && request->seg_cnt() < 32) {
@@ -811,15 +811,15 @@ void TabletImpl::CreateTable(RpcController* controller,
             response->set_msg("table with tid and pid exists");
             done->Run();
             return;
-        }       
+        }
         CreateTableInternal(request, response);
     }
     done->Run();
-    LOG(INFO, "create table with id %d pid %d name %s seg_cnt %d ttl %d", tid, 
+    LOG(INFO, "create table with id %d pid %d name %s seg_cnt %d ttl %lu", tid, 
             pid, name.c_str(), seg_cnt, ttl);
     if (ttl > 0) {
         gc_pool_.DelayTask(FLAGS_gc_interval * 60 * 1000, boost::bind(&TabletImpl::GcTable, this, tid, pid));
-        LOG(INFO, "table %s with tid %ld pid %ld enable ttl %ld", name.c_str(), tid, pid, ttl);
+        LOG(INFO, "table %s with tid %ld pid %ld enable ttl %lu", name.c_str(), tid, pid, ttl);
     }
 }
